+++ sync-basalam/includes/Jobs/Types/UpdateAllProductsJob.php
@@ -27,7 +27,7 @@ class UpdateAllProductsJob extends AbstractJobType
 
     public function getPriority(): int
     {
-        return 2;
+        return 3;
     }
 
     public function canRun(): bool